Fix: Remove old tests
