+++ src/components/Appointment/AvatarGroup.js
@@ -1,13 +1,13 @@
-import { Avatar} from 'flowbite-react/lib/esm/components';
+import { Avatar } from 'flowbite-react/lib/esm/components';
 
-function AvatarGroup({size}){
-
-    return( <Avatar.Group  >
+function AvatarGroup({ size }) {
+  return (
+    <Avatar.Group>
       <Avatar
         img="https://flowbite.com/docs/images/people/profile-picture-1.jpg"
         rounded
         stacked
-       size={size}
+        size={size}
       />
       <Avatar
         img="https://flowbite.com/docs/images/people/profile-picture-2.jpg"
@@ -33,7 +33,8 @@ function AvatarGroup({size}){
         stacked
         size={size}
       />
-    </Avatar.Group>);
-  }
+    </Avatar.Group>
+  );
+}
 
-  export default AvatarGroup;
+export default AvatarGroup;
