+++ src/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { store } from './app/store';
 import App from './App';
+import { store } from './app/store';
 import './index.css';
 
 const container = document.getElementById('root');
