+++ src/pages/Home/Home.js
@@ -1,10 +1,12 @@
 import React from 'react';
-import Appointment from '../../components/appointment/Appointment'
+import Appointment from '../../components/Appointment/Appointment';
 
 function Home() {
-  return (<div className="Home">
-         <Appointment/>
-  </div>);
+  return (
+    <div className="Home">
+      <Appointment />
+    </div>
+  );
 }
 
 export default Home;
